backend/requirements.txt
django
djangorestframework
django-cors-headers
djoser
django-rest-authtoken
Pillow